+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 -------------------
 
 * cost function now takes the same kwargs as process_circuits
+* Updated pytket version requirement to 1.14.1rc0
 
 0.15.0 (April 2023)
 -------------------